simplify TFM consumption
