+++ src/protobuf-net.ServiceModel/Properties/AssemblyInfo.cs
@@ -1,0 +1,6 @@
+﻿
+using System.Runtime.CompilerServices;
+
+#if PLAT_SKIP_LOCALS_INIT
+[module: SkipLocalsInit]
+#endif
